Updated plugins's main file to match with plugin's slug.
